bug fixes, separation of classes into files, standardizing toml names, removing unused files from the project to unused folder, inlcude newest correct yml file
