+++ .gitignore
@@ -3,3 +3,4 @@
 *.cpython*
 visualization_learning
 collector_tests
+unused_files
