--- area_det.yml
@@ -1,186 +1,0 @@
-name: area_det
-channels:
-  - conda-forge
-  - defaults
-dependencies:
-  - appnope=0.1.3=pyhd8ed1ab_0
-  - asttokens=2.4.0=pyhd8ed1ab_0
-  - backcall=0.2.0=pyh9f0ad1d_0
-  - backports=1.0=pyhd8ed1ab_3
-  - backports.functools_lru_cache=1.6.5=pyhd8ed1ab_0
-  - comm=0.1.4=pyhd8ed1ab_0
-  - decorator=5.1.1=pyhd8ed1ab_0
-  - importlib-metadata=6.8.0=pyha770c72_0
-  - importlib_metadata=6.8.0=hd8ed1ab_0
-  - jedi=0.19.1=pyhd8ed1ab_0
-  - jupyter_client=8.5.0=pyhd8ed1ab_0
-  - matplotlib-inline=0.1.6=pyhd8ed1ab_0
-  - nest-asyncio=1.5.8=pyhd8ed1ab_0
-  - packaging=23.2=pyhd8ed1ab_0
-  - parso=0.8.3=pyhd8ed1ab_0
-  - pexpect=4.8.0=pyh1a96a4e_2
-  - pickleshare=0.7.5=py_1003
-  - pip=23.3
-  - platformdirs=3.11.0=pyhd8ed1ab_0
-  - prompt-toolkit=3.0.39=pyha770c72_0
-  - prompt_toolkit=3.0.39=hd8ed1ab_0
-  - ptyprocess=0.7.0=pyhd3deb0d_0
-  - pure_eval=0.2.2=pyhd8ed1ab_0
-  - pygments=2.16.1=pyhd8ed1ab_0
-  - python=3.8.18
-  - python-dateutil=2.8.2=pyhd8ed1ab_0
-  - python_abi=3.8=2_cp38
-  - pyzmq=24.0.1
-  - readline=8.2
-  - setuptools=68.0.0
-  - six=1.16.0=pyh6c4a22f_0
-  - sqlite=3.41.2
-  - stack_data=0.6.2=pyhd8ed1ab_0
-  - tk=8.6.12
-  - tornado=6.3.3
-  - traitlets=5.12.0=pyhd8ed1ab_0
-  - typing-extensions=4.8.0=hd8ed1ab_0
-  - typing_extensions=4.8.0=pyha770c72_0
-  - wcwidth=0.2.8=pyhd8ed1ab_0
-  - wheel=0.41.2
-  - xz=5.4.2
-  - zeromq=4.3.4
-  - zipp=3.17.0=pyhd8ed1ab_0
-  - zlib=1.2.13=h4dc903c_0
-  - pip:
-      - alabaster==0.7.13
-      - appdirs==1.4.4
-      - apstools==1.6.17
-      - area-detector-handlers==0.0.10
-      - asciitree==0.3.3
-      - asteval==0.9.31
-      - attrs==23.1.0
-      - babel==2.13.1
-      - backports-zoneinfo==0.2.1
-      - blosc==1.11.1
-      - bluesky==1.11.0
-      - bluesky-live==0.0.8
-      - boltons==23.0.0
-      - c2dataviewer==1.11.2
-      - cachetools==5.3.2
-      - certifi==2023.7.22
-      - charset-normalizer==3.3.1
-      - click==8.1.7
-      - cloudpickle==3.0.0
-      - contourpy==1.1.1
-      - cycler==0.12.1
-      - dask==2023.5.0
-      - databroker==1.2.5
-      - databroker-pack==0.3.0
-      - dnspython==2.4.2
-      - doct==1.1.0
-      - docutils==0.20.1
-      - entrypoints==0.4
-      - et-xmlfile==1.1.0
-      - event-model==1.19.8
-      - executing==2.0.0
-      - fasteners==0.19
-      - fonttools==4.43.1
-      - fsspec==2023.10.0
-      - future==0.18.3
-      - greenlet==3.0.1
-      - h5py==3.10.0
-      - hdf5plugin==4.2.0
-      - heapdict==1.0.1
-      - historydict==1.2.6
-      - humanize==4.8.0
-      - idna==3.4
-      - imageio==2.31.6
-      - imagesize==1.4.1
-      - importlib-resources==6.1.0
-      - intake==0.6.4
-      - ipython==8.12.3
-      - jinja2==3.1.2
-      - joblib==1.3.2
-      - jsonschema==4.19.1
-      - jsonschema-specifications==2023.7.1
-      - kiwisolver==1.4.5
-      - lazy-loader==0.3
-      - lmfit==1.2.2
-      - locket==1.0.0
-      - lxml==4.9.3
-      - markupsafe==2.1.3
-      - matplotlib==3.7.3
-      - mongoquery==1.4.2
-      - msgpack==1.0.7
-      - msgpack-numpy==0.4.8
-      - networkx==3.1
-      - numcodecs==0.12.1
-      - numpy==1.24.4
-      - numpydoc==1.6.0
-      - openpyxl==3.1.2
-      - ophyd==1.9.0
-      - pandas==2.0.3
-      - partd==1.4.1
-      - pillow==10.0.1
-      - pims==0.6.1
-      - pint==0.21.1
-      - pkgutil-resolve-name==1.3.10
-      - polartools==0.4.0
-      - prettytable==3.9.0
-      - psutil==5.9.6
-      - pvapy==5.3.1
-      - pyepics==3.5.2
-      - pymongo==4.5.0
-      - pyparsing==3.1.1
-      - pyqt5==5.15.10
-      - pyqt5-qt5==5.15.11
-      - pyqt5-sip==12.13.0
-      - pyqtgraph==0.12.4
-      - pyqtwebengine==5.15.6
-      - pyqtwebengine-qt5==5.15.2
-      - pyresttable==2020.0.8
-      - pysumreg==1.0.5
-      - pytz==2023.3.post1
-      - pywavelets==1.4.1
-      - pyyaml==6.0.1
-      - referencing==0.30.2
-      - requests==2.31.0
-      - rpds-py==0.10.6
-      - scikit-image==0.20.0
-      - scikit-learn==1.3.2
-      - scipy==1.9.1
-      - setuptools-scm==8.0.4
-      - slicerator==1.1.0
-      - snowballstemmer==2.2.0
-      - spec2nexus==2021.2.6
-      - sphinx==7.1.2
-      - sphinx-copybutton==0.5.2
-      - sphinxcontrib-applehelp==1.0.4
-      - sphinxcontrib-devhelp==1.0.2
-      - sphinxcontrib-htmlhelp==2.0.1
-      - sphinxcontrib-jsmath==1.0.1
-      - sphinxcontrib-qthelp==1.0.3
-      - sphinxcontrib-serializinghtml==1.1.5
-      - sqlalchemy==2.0.22
-      - stack-data==0.6.3
-      - suitcase-csv==0.3.0
-      - suitcase-json-metadata==0.2.1
-      - suitcase-jsonl==0.2.2
-      - suitcase-mongo==0.4.0
-      - suitcase-msgpack==0.3.0
-      - suitcase-utils==0.5.4
-      - super-state-machine==2.0.2
-      - tabulate==0.9.0
-      - threadpoolctl==3.2.0
-      - tifffile==2023.7.10
-      - tomli==2.0.1
-      - toolz==0.12.0
-      - tqdm==4.66.1
-      - tzdata==2023.3
-      - tzlocal==5.2
-      - uncertainties==3.1.7
-      - urllib3==2.0.7
-      - xarray==2023.1.0
-      - xlrd==2.0.1
-      - xplotutil==1.1.6
-      - xraydb==4.5.1
-      - xrayutilities==1.7.4
-      - xrd-image-util==1.1.3
-      - zarr==2.16.1
-      - zict==2.2.0
