--- environment_py311.yml
@@ -1,186 +1,0 @@
-name: dashpva_py311
-channels:
-  - apsu
-  - conda-forge
-  - defaults
-dependencies:
-  - _libgcc_mutex=0.1=conda_forge
-  - _openmp_mutex=4.5=2_gnu
-  - alsa-lib=1.2.13=hb9d3cd8_0
-  - asteval=1.0.6=pyhd8ed1ab_0
-  - attr=2.5.1=h166bdaf_1
-  - brotli=1.1.0=hb9d3cd8_2
-  - brotli-bin=1.1.0=hb9d3cd8_2
-  - bzip2=1.0.8=h5eee18b_6
-  - c-ares=1.34.4=hb9d3cd8_0
-  - ca-certificates=2025.1.31=hbcca054_0
-  - cached-property=1.5.2=hd8ed1ab_1
-  - cached_property=1.5.2=pyha770c72_1
-  - cairo=1.18.0=h3faef2a_0
-  - contourpy=1.3.1=py311hd18a35c_0
-  - cycler=0.12.1=pyhd8ed1ab_1
-  - dbus=1.13.6=h5008d03_3
-  - dill=0.3.9=pyhd8ed1ab_1
-  - epics-base=7.0.8.1.1.pvapy=pl5262_2
-  - expat=2.6.4=h5888daf_0
-  - font-ttf-dejavu-sans-mono=2.37=hab24e00_0
-  - font-ttf-inconsolata=3.000=h77eed37_0
-  - font-ttf-source-code-pro=2.038=h77eed37_0
-  - font-ttf-ubuntu=0.83=h77eed37_3
-  - fontconfig=2.15.0=h7e30c49_1
-  - fonts-conda-ecosystem=1=0
-  - fonts-conda-forge=1=0
-  - fonttools=4.56.0=py311h2dc5d0c_0
-  - freetype=2.12.1=h267a509_2
-  - future=1.0.0=pyhd8ed1ab_2
-  - gettext=0.23.1=h5888daf_0
-  - gettext-tools=0.23.1=h5888daf_0
-  - glib=2.82.2=h07242d1_1
-  - glib-tools=2.82.2=h4833e2c_1
-  - graphite2=1.3.13=h59595ed_1003
-  - gst-plugins-base=1.24.4=h9ad1361_0
-  - gstreamer=1.24.4=haf2f30d_0
-  - h5py=3.12.1=nompi_py311h5ed33ec_103
-  - harfbuzz=8.5.0=hfac3d4d_0
-  - hdf5=1.14.4=nompi_h2d575fe_105
-  - icu=73.2=h59595ed_0
-  - importlib_resources=6.5.2=pyhd8ed1ab_0
-  - keyutils=1.6.1=h166bdaf_0
-  - kiwisolver=1.4.7=py311hd18a35c_0
-  - krb5=1.21.3=h659f571_0
-  - lame=3.100=h166bdaf_1003
-  - lcms2=2.17=h717163a_0
-  - ld_impl_linux-64=2.40=h12ee557_0
-  - lerc=4.0.0=h27087fc_0
-  - libaec=1.1.3=h59595ed_0
-  - libasprintf=0.23.1=h8e693c7_0
-  - libasprintf-devel=0.23.1=h8e693c7_0
-  - libblas=3.9.0=28_h59b9bed_openblas
-  - libbrotlicommon=1.1.0=hb9d3cd8_2
-  - libbrotlidec=1.1.0=hb9d3cd8_2
-  - libbrotlienc=1.1.0=hb9d3cd8_2
-  - libcap=2.71=h39aace5_0
-  - libcblas=3.9.0=28_he106b2a_openblas
-  - libclang-cpp15=15.0.7=default_h127d8a8_5
-  - libclang13=19.1.2=default_h9c6a7e4_1
-  - libcups=2.3.3=h4637d8d_4
-  - libcurl=8.11.1=h332b0f4_0
-  - libdeflate=1.23=h4ddbbb0_0
-  - libedit=3.1.20191231=he28a2e2_2
-  - libev=4.33=hd590300_2
-  - libevent=2.1.12=hf998b51_1
-  - libexpat=2.6.4=h5888daf_0
-  - libffi=3.4.4=h6a678d5_1
-  - libflac=1.4.3=h59595ed_0
-  - libgcc=14.2.0=h77fa898_1
-  - libgcc-ng=14.2.0=h69a702a_1
-  - libgcrypt-lib=1.11.0=hb9d3cd8_2
-  - libgettextpo=0.23.1=h5888daf_0
-  - libgettextpo-devel=0.23.1=h5888daf_0
-  - libgfortran=14.2.0=h69a702a_1
-  - libgfortran5=14.2.0=hd5240d6_1
-  - libglib=2.82.2=h2ff4ddf_1
-  - libgomp=14.2.0=h77fa898_1
-  - libgpg-error=1.51=hbd13f7d_1
-  - libiconv=1.17=hd590300_2
-  - libjpeg-turbo=3.0.0=hd590300_1
-  - liblapack=3.9.0=28_h7ac8fdf_openblas
-  - libllvm15=15.0.7=hb3ce162_4
-  - libllvm19=19.1.2=ha7bfdaf_0
-  - liblzma=5.6.4=hb9d3cd8_0
-  - libnghttp2=1.64.0=h161d5f1_0
-  - libnsl=2.0.1=hd590300_0
-  - libogg=1.3.5=h4ab18f5_0
-  - libopenblas=0.3.28=pthreads_h94d23a6_1
-  - libopus=1.3.1=h7f98852_1
-  - libpng=1.6.46=h943b412_0
-  - libpq=16.7=h87c4ccc_0
-  - libsndfile=1.2.2=hc60ed4a_1
-  - libsqlite=3.45.2=h2797004_0
-  - libssh2=1.11.1=hf672d98_0
-  - libstdcxx=14.2.0=hc0a3c3a_1
-  - libstdcxx-ng=14.2.0=h4852527_1
-  - libsystemd0=257.2=h3dc2cb9_0
-  - libtiff=4.7.0=hd9ff511_3
-  - libuuid=2.38.1=h0b41bf4_0
-  - libvorbis=1.3.7=h9c3ff4c_0
-  - libwebp-base=1.5.0=h851e524_0
-  - libxcb=1.15=h0b41bf4_0
-  - libxcrypt=4.4.36=hd590300_1
-  - libxkbcommon=1.7.0=h662e7e4_0
-  - libxml2=2.12.7=h4c95cb1_3
-  - libzlib=1.3.1=hb9d3cd8_2
-  - lmfit=1.3.2=pyhd8ed1ab_1
-  - lz4-c=1.10.0=h5888daf_1
-  - matplotlib-base=3.10.0=py311h2b939e6_0
-  - mpg123=1.32.9=hc50e24c_0
-  - munkres=1.1.4=pyh9f0ad1d_0
-  - mysql-common=8.3.0=h70512c7_5
-  - mysql-libs=8.3.0=ha479ceb_5
-  - ncurses=6.4=h6a678d5_0
-  - nspr=4.36=h5888daf_0
-  - nss=3.98=h1d7d5a4_0
-  - numpy=2.2.3=py311h5d046bc_0
-  - openjpeg=2.5.3=h5fbd93e_0
-  - openssl=3.4.1=h7b32b05_0
-  - packaging=24.2=pyhd8ed1ab_2
-  - pcre2=10.44=hba22ea6_2
-  - perl=5.32.1=7_hd590300_perl5
-  - pillow=10.3.0=py311h18e6fac_0
-  - pip=25.0=py311h06a4308_0
-  - pixman=0.44.2=h29eaf8c_0
-  - ply=3.11=pyhd8ed1ab_3
-  - pthread-stubs=0.4=hb9d3cd8_1002
-  - pulseaudio-client=17.0=hb77b528_0
-  - pvapy=5.5.0=py311_1
-  - pvapy-boost=1.87.0=py311_1
-  - pyepics=3.5.7=py311h38be061_1
-  - pyparsing=3.2.1=pyhd8ed1ab_0
-  - pyqt=5.15.9=py311hf0fb5b6_5
-  - pyqt5-sip=12.12.2=py311hb755f60_5
-  - pyqtgraph=0.13.7=pyhd8ed1ab_1
-  - python=3.11.8=hab00c5b_0_cpython
-  - python-dateutil=2.9.0.post0=pyhff2d567_1
-  - python_abi=3.11=5_cp311
-  - qhull=2020.2=h434a139_5
-  - qt-main=5.15.8=hc9dc06e_21
-  - readline=8.2=h5eee18b_0
-  - scipy=1.15.1=py311hc1ac118_0
-  - setuptools=75.8.0=py311h06a4308_0
-  - setuptools-scm=8.1.0=pyhd8ed1ab_1
-  - sip=6.7.12=py311hb755f60_0
-  - six=1.17.0=pyhd8ed1ab_0
-  - sqlite=3.45.2=h2c6b66d_0
-  - tk=8.6.13=noxft_h4845f30_101
-  - toml=0.10.2=pyhd8ed1ab_1
-  - tomli=2.2.1=pyhd8ed1ab_1
-  - typing-extensions=4.12.2=hd8ed1ab_1
-  - typing_extensions=4.12.2=pyha770c72_1
-  - tzdata=2025a=h04d1e81_0
-  - uncertainties=3.2.2=pyhd8ed1ab_2
-  - unicodedata2=16.0.0=py311h9ecbd09_0
-  - wheel=0.45.1=py311h06a4308_0
-  - xcb-util=0.4.0=hd590300_1
-  - xcb-util-image=0.4.0=h8ee46fc_1
-  - xcb-util-keysyms=0.4.0=h8ee46fc_1
-  - xcb-util-renderutil=0.3.9=hd590300_1
-  - xcb-util-wm=0.4.1=h8ee46fc_1
-  - xkeyboard-config=2.42=h4ab18f5_0
-  - xorg-kbproto=1.0.7=hb9d3cd8_1003
-  - xorg-libice=1.1.2=hb9d3cd8_0
-  - xorg-libsm=1.2.5=he73a12e_0
-  - xorg-libx11=1.8.9=h8ee46fc_0
-  - xorg-libxau=1.0.12=hb9d3cd8_0
-  - xorg-libxdmcp=1.1.5=hb9d3cd8_0
-  - xorg-libxext=1.3.4=h0b41bf4_2
-  - xorg-libxrender=0.9.11=hd590300_0
-  - xorg-renderproto=0.11.1=hb9d3cd8_1003
-  - xorg-xextproto=7.3.0=hb9d3cd8_1004
-  - xorg-xf86vidmodeproto=2.3.1=hb9d3cd8_1005
-  - xorg-xproto=7.0.31=hb9d3cd8_1008
-  - xrayutilities=1.7.10=py311h9f3472d_0
-  - xz=5.6.4=h5eee18b_1
-  - zipp=3.21.0=pyhd8ed1ab_1
-  - zlib=1.3.1=hb9d3cd8_2
-  - zstd=1.5.6=ha6fb4c9_0
-prefix: /home/beams/JULIO.RODRIGUEZ/anaconda3/envs/area_det_py311
