--- environment_py37.yml
@@ -1,196 +1,0 @@
-name: dashpva_py37
-channels:
-  - epics
-  - conda-forge
-  - defaults
-dependencies:
-  - _libgcc_mutex=0.1=main
-  - _openmp_mutex=4.5=1_gnu
-  - anyio=3.7.1=pyhd8ed1ab_0
-  - argon2-cffi=23.1.0=pyhd8ed1ab_0
-  - argon2-cffi-bindings=21.2.0=py37h7f8727e_0
-  - backcall=0.2.0=pyh9f0ad1d_0
-  - backports=1.0=pyhd8ed1ab_3
-  - backports.functools_lru_cache=2.0.0=pyhd8ed1ab_0
-  - beautifulsoup4=4.12.3=pyha770c72_0
-  - blas=1.0=mkl
-  - bleach=6.1.0=pyhd8ed1ab_0
-  - brotli=1.0.9=he6710b0_2
-  - ca-certificates=2024.12.31=h06a4308_0
-  - certifi=2024.2.2=pyhd8ed1ab_0
-  - cudatoolkit=10.2.89=h8f6ccaa_8
-  - cupy=10.1.0=py37ha5d5e86_0
-  - cycler=0.11.0=pyhd3eb1b0_0
-  - dbus=1.13.18=hb2f20db_0
-  - debugpy=1.4.1=py37hcd2ae1e_0
-  - decorator=5.1.1=pyhd8ed1ab_0
-  - defusedxml=0.7.1=pyhd8ed1ab_0
-  - entrypoints=0.4=pyhd8ed1ab_0
-  - epics-base=7.0.7=pl5262_2
-  - exceptiongroup=1.2.0=pyhd8ed1ab_2
-  - expat=2.4.4=h295c915_0
-  - fastrlock=0.6=py37hcd2ae1e_1
-  - fontconfig=2.13.1=h6c09931_0
-  - fonttools=4.25.0=pyhd3eb1b0_0
-  - freetype=2.11.0=h70c0345_0
-  - giflib=5.2.1=h7b6447c_0
-  - glib=2.69.1=h4ff587b_1
-  - gst-plugins-base=1.14.0=h8213a91_2
-  - gstreamer=1.14.0=h28cd5cc_2
-  - icu=58.2=he6710b0_3
-  - importlib-metadata=4.11.4=py37h89c1867_0
-  - importlib_resources=6.0.0=pyhd8ed1ab_0
-  - intel-openmp=2021.4.0=h06a4308_3561
-  - ipykernel=6.16.2=pyh210e3f2_0
-  - ipython=7.33.0=py37h89c1867_0
-  - ipython_genutils=0.2.0=py_1
-  - jedi=0.19.1=pyhd8ed1ab_0
-  - jinja2=3.1.3=pyhd8ed1ab_0
-  - jpeg=9e=h7f8727e_0
-  - jsonschema=4.17.3=pyhd8ed1ab_0
-  - jupyter_client=7.1.2=pyhd8ed1ab_0
-  - jupyter_core=4.11.1=py37h89c1867_0
-  - jupyter_server=1.23.4=pyhd8ed1ab_0
-  - jupyterlab_pygments=0.3.0=pyhd8ed1ab_1
-  - kiwisolver=1.3.2=py37h295c915_0
-  - lcms2=2.12=h3be6417_0
-  - ld_impl_linux-64=2.35.1=h7274673_9
-  - libffi=3.3=he6710b0_2
-  - libgcc-ng=9.3.0=h5101ec6_17
-  - libgomp=9.3.0=h5101ec6_17
-  - libpng=1.6.37=hbc83047_0
-  - libsodium=1.0.18=h36c2ea0_1
-  - libstdcxx-ng=9.3.0=hd4cf53a_17
-  - libtiff=4.2.0=h85742a9_0
-  - libuuid=1.0.3=h7f8727e_2
-  - libwebp=1.2.2=h55f646e_0
-  - libwebp-base=1.2.2=h7f8727e_0
-  - libxcb=1.14=h7b6447c_0
-  - libxml2=2.9.12=h74e7548_1
-  - lz4-c=1.9.3=h295c915_1
-  - markupsafe=2.0.1=py37h5e8e339_0
-  - matplotlib=3.4.2=py37h06a4308_0
-  - matplotlib-base=3.4.2=py37hab158f2_0
-  - matplotlib-inline=0.1.6=pyhd8ed1ab_0
-  - mistune=3.0.2=pyhd8ed1ab_0
-  - mkl=2021.4.0=h06a4308_640
-  - mkl-service=2.4.0=py37h7f8727e_0
-  - mkl_fft=1.3.1=py37hd3c417c_0
-  - mkl_random=1.2.2=py37h51133e4_0
-  - munkres=1.1.4=py_0
-  - nbclassic=1.0.0=pyhb4ecaf3_1
-  - nbclient=0.7.0=pyhd8ed1ab_0
-  - nbconvert=7.6.0=pyhd8ed1ab_0
-  - nbconvert-core=7.6.0=pyhd8ed1ab_0
-  - nbconvert-pandoc=7.6.0=pyhd8ed1ab_0
-  - nbformat=5.8.0=pyhd8ed1ab_0
-  - ncurses=6.3=h7f8727e_2
-  - nest-asyncio=1.6.0=pyhd8ed1ab_0
-  - notebook=6.5.6=pyha770c72_0
-  - notebook-shim=0.2.4=pyhd8ed1ab_0
-  - numpy=1.20.3=py37hf144106_0
-  - numpy-base=1.20.3=py37h74d4b33_0
-  - openssl=1.1.1w=h7f8727e_0
-  - packaging=23.2=pyhd8ed1ab_0
-  - pandoc=3.1.12.2=ha770c72_0
-  - pandocfilters=1.5.0=pyhd8ed1ab_0
-  - parso=0.8.3=pyhd8ed1ab_0
-  - pcre=8.45=h295c915_0
-  - perl=5.26.2=h14c3975_0
-  - pexpect=4.9.0=pyhd8ed1ab_0
-  - pickleshare=0.7.5=py_1003
-  - pillow=9.0.1=py37h22f2fdc_0
-  - pip=21.2.2=py37h06a4308_0
-  - pkgutil-resolve-name=1.3.10=pyhd8ed1ab_1
-  - prometheus_client=0.17.1=pyhd8ed1ab_0
-  - prompt-toolkit=3.0.42=pyha770c72_0
-  - psutil=5.8.0=py37h5e8e339_1
-  - ptyprocess=0.7.0=pyhd3deb0d_0
-  - pvapy=5.3.1=py37_1
-  - pvapy-boost=1.81.0=py37_2
-  - pycparser=2.21=pyhd8ed1ab_0
-  - pygments=2.17.2=pyhd8ed1ab_0
-  - pyparsing=3.0.4=pyhd3eb1b0_0
-  - pyqt=5.9.2=py37h05f1152_2
-  - pyrsistent=0.17.3=py37h5e8e339_2
-  - python=3.7.13=h12debd9_0
-  - python-dateutil=2.8.2=pyhd3eb1b0_0
-  - python-fastjsonschema=2.19.1=pyhd8ed1ab_0
-  - python_abi=3.7=2_cp37m
-  - pyzmq=22.1.0=py37h336d617_0
-  - qt=5.9.7=h5867ecd_1
-  - readline=8.1.2=h7f8727e_1
-  - send2trash=1.8.2=pyh41d4057_0
-  - sip=4.19.8=py37hf484d3e_0
-  - six=1.16.0=pyhd3eb1b0_1
-  - sniffio=1.3.1=pyhd8ed1ab_0
-  - soupsieve=2.3.2.post1=pyhd8ed1ab_0
-  - sqlite=3.38.3=hc218d9a_0
-  - terminado=0.17.1=pyh41d4057_0
-  - tinycss2=1.2.1=pyhd8ed1ab_0
-  - tk=8.6.11=h1ccaba5_1
-  - toml=0.10.2=pyhd3eb1b0_0
-  - tornado=6.1=py37h27cfd23_0
-  - traitlets=5.9.0=pyhd8ed1ab_0
-  - typing_extensions=4.7.1=pyha770c72_0
-  - wcwidth=0.2.10=pyhd8ed1ab_0
-  - webencodings=0.5.1=pyhd8ed1ab_2
-  - websocket-client=1.6.1=pyhd8ed1ab_0
-  - wheel=0.37.1=pyhd3eb1b0_0
-  - xz=5.2.5=h7f8727e_1
-  - zeromq=4.3.4=h9c3ff4c_0
-  - zipp=3.15.0=pyhd8ed1ab_0
-  - zlib=1.2.12=h7f8727e_2
-  - zstd=1.4.9=haebb681_0
-  - pip:
-      - aiohttp==3.8.1
-      - aiosignal==1.2.0
-      - async-timeout==4.0.2
-      - asynctest==0.13.0
-      - attrs==21.4.0
-      - autobahn==22.4.2
-      - automat==20.2.0
-      - cached-property==1.5.2
-      - cffi==1.15.0
-      - charset-normalizer==2.0.12
-      - click==8.1.7
-      - constantly==15.1.0
-      - cryptography==37.0.2
-      - docopt==0.6.2
-      - frozenlist==1.3.0
-      - h5py==3.6.0
-      - hyperlink==21.0.0
-      - idna==3.3
-      - image-analysis==0.1.5
-      - incremental==21.3.0
-      - joblib==1.1.0
-      - lxml==4.8.0
-      - multidict==6.0.2
-      - pyepics==3.5.1
-      - pyqt5==5.15.9
-      - pyqt5-plugins==5.15.9.2.3
-      - pyqt5-qt5==5.15.2
-      - pyqt5-sip==12.13.0
-      - pyqt5-tools==5.15.9.3.3
-      - pyqtchart==5.15.6
-      - pyqtchart-qt5==5.15.14
-      - pyqtgraph==0.12.4
-      - pyresttable==2020.0.3
-      - python-dotenv==0.21.1
-      - qt5-applications==5.15.2.2.3
-      - qt5-tools==5.15.2.1.3
-      - rsmap3d==1.2.1
-      - scikit-learn==1.0.2
-      - scipy==1.6.3
-      - setuptools==56.2.0
-      - spec2nexus==2021.1.11
-      - threadpoolctl==3.1.0
-      - tifffile==2021.4.8
-      - twisted==22.4.0
-      - txaio==22.2.1
-      - typing-extensions==4.2.0
-      - vtk==9.0.3
-      - wslink==1.6.4
-      - xrayutilities==1.7.1
-      - yarl==1.7.2
-      - zope-interface==5.4.0
